Refactor LanguageFilter and extract shared param util
LanguageFilter.tsx:
- Import createMemo and wrap nativeNames, uiNames, and languages() in
  createMemo so that Intl.DisplayNames instances are created at most
  once per locale change rather than on every reactive access.
- Change nativeNames to use i18n.locale (consistent with uiNames)
  instead of undefined, so the fallback display name matches the
  current UI language.

languageParam.ts (new):
- Extract normalizeLanguageParam into a shared utility at
  web-next/src/lib/languageParam.ts and import it in the five route
  files that previously each defined their own copy of the function.

https://github.com/hackers-pub/hackerspub/pull/302#discussion_r3266009903
https://github.com/hackers-pub/hackerspub/pull/302#discussion_r3266009913
https://github.com/hackers-pub/hackerspub/pull/302#discussion_r3266009917
https://github.com/hackers-pub/hackerspub/pull/302#discussion_r3266009927
