+++ web-next/src/components/LanguageFilter.tsx
@@ -1,5 +1,5 @@
 import { A } from "@solidjs/router";
-import { For, Show } from "solid-js";
+import { createMemo, For, Show } from "solid-js";
 import { useLingui } from "~/lib/i18n/macro.d.ts";
 
 export interface LanguageFilterProps {
@@ -12,20 +12,21 @@ export function LanguageFilter(props: LanguageFilterProps) {
   const { t, i18n } = useLingui();
 
   // Display name in the language itself ("한국어", "日本語", "English")
-  const nativeNames = () => {
+  const nativeNames = createMemo(() => {
     try {
-      return new Intl.DisplayNames(undefined, {
+      return new Intl.DisplayNames(i18n.locale, {
         type: "language",
         fallback: "code",
       });
     } catch {
       return null;
     }
-  };
+  });
 
   // Display name in the current UI locale ("Korean", "日本語", "英語")
-  const uiNames = () =>
-    new Intl.DisplayNames(i18n.locale, { type: "language", fallback: "code" });
+  const uiNames = createMemo(() =>
+    new Intl.DisplayNames(i18n.locale, { type: "language", fallback: "code" })
+  );
 
   const pillClass = (active: boolean) =>
     [
@@ -37,13 +38,13 @@ export function LanguageFilter(props: LanguageFilterProps) {
 
   // Ensure the active language is always shown even if it isn't in the
   // suggested list (e.g. manually typed in the URL).
-  const languages = () => {
+  const languages = createMemo(() => {
     const active = props.activeLanguage;
     if (active == null || props.languages.includes(active)) {
       return props.languages;
     }
     return [active, ...props.languages];
-  };
+  });
 
   return (
     <div class="flex flex-wrap gap-2 border-b px-4 py-3">
