+++ web-next/src/routes/(root)/feed/articles.tsx
@@ -17,6 +17,7 @@ import { PersonalTimeline } from "~/components/PersonalTimeline.tsx";
 import { useViewer } from "~/contexts/ViewerContext.tsx";
 import { buildSignInHref, gateOnAuthentication } from "~/lib/authGate.ts";
 import { useLingui } from "~/lib/i18n/macro.d.ts";
+import { normalizeLanguageParam } from "~/lib/languageParam.ts";
 import { routePreloadedQuery } from "~/lib/relayPreload.ts";
 import type { articlesFeedTimelineQuery } from "./__generated__/articlesFeedTimelineQuery.graphql.ts";
 
@@ -53,18 +54,6 @@ const loadArticlesFeedTimelineQuery = routePreloadedQuery(
     ),
   "loadArticlesFeedTimelineQuery",
 );
-
-function normalizeLanguageParam(
-  raw: string | string[] | undefined,
-): string | undefined {
-  const tag = Array.isArray(raw) ? raw[0] : raw;
-  if (!tag) return undefined;
-  try {
-    return new Intl.Locale(tag).language;
-  } catch {
-    return undefined;
-  }
-}
 
 function AuthenticatedArticlesFeedTimeline() {
   const { i18n } = useLingui();
